refactoring and consolidation of reusable code
