+++ rocks.kfs.Intacct/IntacctOtherReceipt.cs
@@ -25,6 +25,7 @@ using Rock.Data;
 using Rock.Model;
 using Rock.Web.Cache;
 using rocks.kfs.Intacct.Enums;
+using rocks.kfs.Intacct.Utils;
 using KFSConst = rocks.kfs.Intacct.SystemGuid;
 
 namespace rocks.kfs.Intacct
@@ -219,159 +220,38 @@ namespace rocks.kfs.Intacct
                 RefId = financialBatch.Id.ToString(),
                 ReceiptItems = new List<ReceiptLineItem>()
             };
-
-            //
-            // Group/Sum Transactions by Debit/Bank Account and Project since Project can come from Account or Transaction Details
-            //
-            var batchTransactions = new List<OtherReceiptTransaction>();
-            var registrationLinks = new List<RegistrationInstance>();
-            var groupMemberLinks = new List<GroupMember>();
-            foreach ( var transaction in financialBatch.Transactions )
-            {
-                transaction.LoadAttributes();
-                var gateway = transaction.FinancialGateway;
-                var gatewayDefaultFeeAccount = string.Empty;
-                var processTransactionFees = 0;
-                if ( gateway != null )
-                {
-                    gateway.LoadAttributes();
-                    gatewayDefaultFeeAccount = transaction.FinancialGateway.GetAttributeValue( "rocks.kfs.Intacct.DEFAULTFEEACCOUNTNO" );
-                    var gatewayFeeProcessing = transaction.FinancialGateway.GetAttributeValue( "rocks.kfs.Intacct.FEEPROCESSING" ).AsIntegerOrNull();
-                    if ( gatewayFeeProcessing != null )
-                    {
-                        processTransactionFees = gatewayFeeProcessing.Value;
-                    }
-                }
-
-                foreach ( var transactionDetail in transaction.TransactionDetails )
-                {
-                    transactionDetail.LoadAttributes();
-                    transactionDetail.Account.LoadAttributes();
-
-                    var detailProject = transactionDetail.GetAttributeValue( "rocks.kfs.Intacct.PROJECTID" ).AsGuidOrNull();
-                    var accountProject = transactionDetail.Account.GetAttributeValue( "rocks.kfs.Intacct.PROJECTID" ).AsGuidOrNull();
-                    var transactionFeeAccount = transactionDetail.Account.GetAttributeValue( "rocks.kfs.Intacct.FEEACCOUNTNO" );
-
-                    if ( string.IsNullOrWhiteSpace( transactionFeeAccount ) )
-                    {
-                        transactionFeeAccount = gatewayDefaultFeeAccount;
-                    }
-
-                    var projectCode = string.Empty;
-                    if ( detailProject != null )
-                    {
-                        projectCode = DefinedValueCache.Get( ( Guid ) detailProject ).Value;
-                    }
-                    else if ( accountProject != null )
-                    {
-                        projectCode = DefinedValueCache.Get( ( Guid ) accountProject ).Value;
-                    }
-
-                    if ( transactionDetail.EntityTypeId.HasValue )
-                    {
-                        var registrationEntityType = EntityTypeCache.Get( typeof( Rock.Model.Registration ) );
-                        var groupMemberEntityType = EntityTypeCache.Get( typeof( Rock.Model.GroupMember ) );
-
-                        if ( transactionDetail.EntityId.HasValue && transactionDetail.EntityTypeId == registrationEntityType.Id )
-                        {
-                            foreach ( var registration in new RegistrationService( rockContext )
-                                .Queryable().AsNoTracking()
-                                .Where( r =>
-                                    r.RegistrationInstance != null &&
-                                    r.RegistrationInstance.RegistrationTemplate != null &&
-                                    r.Id == transactionDetail.EntityId ) )
-                            {
-                                registrationLinks.Add( registration.RegistrationInstance );
-                            }
-                        }
-                        if ( transactionDetail.EntityId.HasValue && transactionDetail.EntityTypeId == groupMemberEntityType.Id )
-                        {
-                            foreach ( var groupMember in new GroupMemberService( rockContext )
-                                .Queryable().AsNoTracking()
-                                .Where( gm =>
-                                    gm.Group != null &&
-                                    gm.Id == transactionDetail.EntityId ) )
-                            {
-                                groupMemberLinks.Add( groupMember );
-                            }
-                        }
-                    }
-
-                    var transactionItem = new OtherReceiptTransaction()
-                    {
-                        TransactionId = transactionDetail.TransactionId,
-                        Payer = transaction.AuthorizedPersonAlias.Person.FullName,
-                        Amount = transactionDetail.Amount,
-                        FinancialAccountId = transactionDetail.AccountId,
-                        Project = projectCode,
-                        PaymentDate = transaction.TransactionDateTime.Value,
-                        TransactionFeeAmount = transactionDetail.FeeAmount != null && transactionDetail.FeeAmount.Value > 0 ? transactionDetail.FeeAmount.Value : 0.0M,
-                        TransactionFeeAccount = transactionFeeAccount,
-                        ProcessTransactionFees = processTransactionFees
-                    };
-
-                    batchTransactions.Add( transactionItem );
-                }
-            }
-
-            var receiptTransactions = batchTransactions
-            .GroupBy( d => new { d.FinancialAccountId, d.Project, d.TransactionFeeAccount, d.ProcessTransactionFees } )
-            .Select( s => new OtherReceiptTransaction
-            {
-                Payer = "Rock Import",
-                TransactionId = -1,
-                FinancialAccountId = s.Key.FinancialAccountId,
-                Project = s.Key.Project,
-                Amount = s.Sum( f => ( decimal? ) f.Amount ) ?? 0.0M,
-                TransactionFeeAmount = s.Sum( f => ( decimal? ) f.TransactionFeeAmount ) ?? 0.0M,
-                TransactionFeeAccount = s.Key.TransactionFeeAccount,
-                ProcessTransactionFees = s.Key.ProcessTransactionFees
-            } )
-            .ToList();
+            List<RegistrationInstance> registrationLinks;
+            List<GroupMember> groupMemberLinks;
+            var receiptTransactions = TransactionHelpers.GetTransactionSummary( financialBatch, rockContext, out registrationLinks, out groupMemberLinks );
 
             //
             // Get the Dimensions from the Account since the Transaction Details have been Grouped already
             //
-            var customDimensions = GetCustomDimensions();
+            var customDimensions = TransactionHelpers.GetCustomDimensions();
 
             // Create Receipt Item for each entry within a grouping
             foreach ( var bTran in receiptTransactions )
             {
                 var account = new FinancialAccountService( rockContext ).Get( bTran.FinancialAccountId );
                 var customDimensionValues = new Dictionary<string, dynamic>();
-                var mergeFields = new Dictionary<string, object>();
-                account.LoadAttributes();
-
-                if ( customDimensions.Count > 0 )
+                var mergeFieldObjects = new MergeFieldObjects
                 {
-                    foreach ( var rockKey in customDimensions )
-                    {
-                        var dimension = rockKey.Split( '.' ).Last();
-                        customDimensionValues.Add( dimension, account.GetAttributeValue( rockKey ) );
-                    }
-                }
+                    Account = account,
+                    Batch = financialBatch,
+                    Registrations = registrationLinks,
+                    GroupMembers = groupMemberLinks,
+                    Summary = bTran,
+                    CustomDimensions = customDimensions
+                };
+                Dictionary<string, object> mergeFields = TransactionHelpers.GetMergeFieldsAndDimensions( ref debugLava, customDimensionValues, mergeFieldObjects );
 
-                mergeFields.Add( "Account", account );
-                mergeFields.Add( "Batch", financialBatch );
-                mergeFields.Add( "Registrations", registrationLinks );
-                mergeFields.Add( "GroupMembers", groupMemberLinks );
-                mergeFields.Add( "CustomDimensions", customDimensionValues );
-                mergeFields.Add( "Summary", bTran );
-
-                if ( debugLava.Length < 6 && debugLava.AsBoolean() )
-                {
-                    debugLava = mergeFields.lavaDebugInfo();
-                }
-
-                var bankAccount = account.GetAttributeValue( "rocks.kfs.Intacct.DEBITACCOUNTNO" );
-                var glAccount = account.GetAttributeValue( "rocks.kfs.Intacct.ACCOUNTNO" );
                 var classId = account.GetAttributeValue( "rocks.kfs.Intacct.CLASSID" );
                 var departmentId = account.GetAttributeValue( "rocks.kfs.Intacct.DEPARTMENT" );
                 var locationId = account.GetAttributeValue( "rocks.kfs.Intacct.LOCATION" );
 
                 var receiptItem = new ReceiptLineItem
                 {
-                    GlAccountNo = glAccount,
+                    GlAccountNo = account.GetAttributeValue( "rocks.kfs.Intacct.ACCOUNTNO" ),
                     Amount = bTran.ProcessTransactionFees == 1 ? bTran.Amount - bTran.TransactionFeeAmount : bTran.Amount,
                     Memo = DescriptionLava.ResolveMergeFields( mergeFields ),
                     LocationId = locationId,
@@ -399,37 +279,6 @@ namespace rocks.kfs.Intacct
             }
 
             return otherReceipt;
-        }
-
-        private List<string> GetCustomDimensions()
-        {
-            var knownDimensions = new List<string>();
-            knownDimensions.Add( "rocks.kfs.Intacct.ACCOUNTNO" );
-            knownDimensions.Add( "rocks.kfs.Intacct.DEBITACCOUNTNO" );
-            knownDimensions.Add( "rocks.kfs.Intacct.FEEACCOUNTNO" );
-            knownDimensions.Add( "rocks.kfs.Intacct.DEFAULTFEEACCOUNTNO" );
-            knownDimensions.Add( "rocks.kfs.Intacct.FEEPROCESSING" );
-            knownDimensions.Add( "rocks.kfs.Intacct.CLASSID" );
-            knownDimensions.Add( "rocks.kfs.Intacct.DEPARTMENT" );
-            knownDimensions.Add( "rocks.kfs.Intacct.LOCATION" );
-            knownDimensions.Add( "rocks.kfs.Intacct.PROJECTID" );
-
-            var rockContext = new RockContext();
-            var accountCategoryId = new CategoryService( rockContext ).Queryable().FirstOrDefault( c => c.Guid.Equals( new System.Guid( KFSConst.Attribute.FINANCIAL_ACCOUNT_ATTRIBUTE_CATEGORY ) ) ).Id;
-            var gatewayCategoryId = new CategoryService( rockContext ).Queryable().FirstOrDefault( c => c.Guid.Equals( new System.Guid( KFSConst.Attribute.FINANCIAL_GATEWAY_ATTRIBUTE_CATEGORY ) ) ).Id;
-            var attributeService = new AttributeService( rockContext );
-            var accountAttributes = attributeService.Queryable().Where( a => a.Categories.Select( c => c.Id ).Contains( accountCategoryId ) ).ToList();
-
-            var customDimensions = new List<string>();
-
-            foreach ( var attribute in accountAttributes )
-            {
-                if ( !knownDimensions.Contains( attribute.Key ) )
-                {
-                    customDimensions.Add( attribute.Key );
-                }
-            }
-            return customDimensions;
         }
     }
 }
