+++ rocks.kfs.Intacct/rocks.kfs.Intacct.csproj
@@ -83,6 +83,7 @@
     <Compile Include="IntacctEndpoint.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SystemGuid\Attribute.cs" />
+    <Compile Include="Utils\TransactionHelpers.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
